requirements.txt
requests
docker
dnspython
